+++ tests/test_enum.py
@@ -143,6 +143,16 @@ def test_enum_undefined_value():
     Test.FieldTest().EnumField = Test.ShortEnum(20, True)
 
 
+def test_enum_repr():
+    """Test enumeration repr."""
+    from System import DayOfWeek
+
+    assert repr(DayOfWeek.Monday) == "<DayOfWeek.Monday: 1>"
+
+    assert repr(Test.FlagsEnum(7)) == "<FlagsEnum.Two, Five: 0x00000007>"
+    assert repr(Test.FlagsEnum(8)) == "<FlagsEnum.8: 0x00000008>"
+
+
 def test_enum_conversion():
     """Test enumeration conversion."""
     ob = Test.FieldTest()